refactor(frontend): Add sse-helper
