+++ frontend/src/app/assignment/services/selection.service.ts
@@ -1,8 +1,9 @@
 import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs';
-import {environment} from '../../../../environments/environment';
-import {Snippet} from '../../model/evaluation';
-import {AssignmentService} from '../../services/assignment.service';
+import {environment} from '../../../environments/environment';
+import {Snippet} from '../model/evaluation';
+import {AssignmentService} from './assignment.service';
+import {observeSSE} from './sse-helper';
 
 interface SelectionDto {
   assignment: string;
@@ -23,11 +24,6 @@ export class SelectionService {
   stream(assignment: string, solution?: string): Observable<{ event: string, selection: SelectionDto }> {
     const token = this.assignmentService.getToken(assignment);
     const url = `${environment.assignmentsApiUrl}/assignments/${assignment}${solution ? '/solutions/' + solution : ''}/selections/events?token=${token}`;
-    return new Observable(observer => {
-      const eventSource = new EventSource(url);
-      eventSource.addEventListener('message', event => observer.next(JSON.parse(event.data)));
-      eventSource.addEventListener('error', error => observer.error(error));
-      return () => eventSource.close();
-    });
+    return observeSSE<SelectionDto, 'selection'>(url);
   }
 }
